requirements.txt
discord.py
raygun4py